pyproject.toml
[tool.poetry]
name = "aiextractor"
version = "0.1.9"
description = ""
authors = ["mjaquier <mjaquier@mjaquier.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.2"
aihelper = "^0.2.9"
pytest-cov = "^2.10.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
extract = 'aiextractor:main'